Respect the scaling factor on Windows 10
Recent Windows versions allow to change the "Scale" factor of the
display. This is particularly useful on high DPI devices, where
displaying windows in the regular pixel size would make it _really_ hard
to read (think: 3240x2160 resolution on a regular laptop, where the
Scale factor would default to 200%).

To maintain compatibility with scale-unaware software, this is an opt-in
feature, where applications have to specify via their manifest that they
are aware of it. Otherwise, Windows will pretend that the resolution is
actually lower.

However, this leads to a problem where calling `screen.capture()`
_thinks_ that the resolution on above-mentioned laptop is actually
1620x1080, yet when capturing a 1620X1080 rectangle at 0x0, it only
captures the upper left quadrant!

Let's capture the full screen and downsize it using the current Scale
factor.

Note: an alternative would be to report the high resolution as desktop
size. However, this is incompatible with `screen.highlight()` and
`mouse.move()` which are unaware of the scale factor. The easier
direction, therefore, seems to be the chosen one: just downscale the
high-resolution screen capture and continue to work in the scale-unaware
mode.

Note, too: yet _another_ alternative would be to change the manifest.
But I think that would not work because it would be `node.exe` whose
manifest would have to change, this Node module's `.dll` cannot change
the application's behavior.

For now, we do not bother trying to fall back to using
`GetScaleFactorForDevice()` on Windows 8 (`GetScaleFactorForMonitor()`
is only available in Windows 8.1 and newer).

To allow _running_ on Windows 8 (or older), we only try to load that
function at runtime rather than linking to it (because the `.dll` would
otherwise not even load on Windows 8).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>